Avoid sorting the fragments
Sorting the fragments was confusing as it was very difficult to reason
on why a certain fragment was packed before another.  Better that that
are packed in the order they appear in the recipes.  Removing the
sorting also highlight a bug in the avoid_tags filter.  The filter did
not take into account the avoid_tags of already packed fragments, it
only used the avoid_tags of the incoming fragment.